requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
django-filter==2.3.0
django-graphql-jwt==0.3.1
django==3.1
graphene-django==2.13.0
graphene-file-upload==1.2.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
pillow==7.2.0
promise==2.3
pyjwt==1.7.1
pytz==2020.1
rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
unidecode==1.1.1
